# ci(benchmarks): add/enable startup benchmark to run in ci