--- a/.gitlab/benchmarks/microbenchmarks.yml
+++ b/.gitlab/benchmarks/microbenchmarks.yml
@@ -59,6 +59,7 @@ microbenchmarks:
         - "http_propagation_inject"
         - "rate_limiter"
         - "packages_update_imported_dependencies"
+        - "startup"
 
 benchmarks-pr-comment:
   image: $MICROBENCHMARKS_CI_IMAGE