--- a/benchmarks/startup/config.yaml
+++ b/benchmarks/startup/config.yaml
@@ -1,0 +1,39 @@
+baseline: &defaults
+  env: {}
+  ddtrace_run: false
+  import_ddtrace: false
+  import_ddtrace_auto: false
+  import_flask: false
+baseline_flask:
+  <<: *defaults
+  import_flask: true
+ddtrace_run:
+  <<: *defaults
+  ddtrace_run: true
+ddtrace_run_flask:
+  <<: *defaults
+  ddtrace_run: true
+  import_flask: true
+import_ddtrace:
+  <<: *defaults
+  import_ddtrace: true
+import_ddtrace_flask:
+  <<: *defaults
+  import_ddtrace: true
+  import_flask: true
+import_ddtrace_auto:
+  <<: *defaults
+  import_ddtrace_auto: true
+import_ddtrace_auto_flask:
+  <<: *defaults
+  import_ddtrace_auto: true
+  import_flask: true
+ddtrace_run_import_ddtrace_auto:
+  <<: *defaults
+  ddtrace_run: true
+  import_ddtrace_auto: true
+ddtrace_run_import_ddtrace_auto_flask:
+  <<: *defaults
+  ddtrace_run: true
+  import_ddtrace_auto: true
+  import_flask: true