--- a/benchmarks/startup/config.yaml
+++ b/benchmarks/startup/config.yaml
@@ -4,9 +4,13 @@ baseline: &defaults
   import_ddtrace: false
   import_ddtrace_auto: false
   import_flask: false
+  import_django: false
 baseline_flask:
   <<: *defaults
   import_flask: true
+baseline_django:
+  <<: *defaults
+  import_django: true
 ddtrace_run:
   <<: *defaults
   ddtrace_run: true
@@ -14,6 +18,10 @@ ddtrace_run_flask:
   <<: *defaults
   ddtrace_run: true
   import_flask: true
+ddtrace_run_django:
+  <<: *defaults
+  ddtrace_run: true
+  import_django: true
 import_ddtrace:
   <<: *defaults
   import_ddtrace: true
@@ -21,6 +29,10 @@ import_ddtrace_flask:
   <<: *defaults
   import_ddtrace: true
   import_flask: true
+import_ddtrace_django:
+  <<: *defaults
+  import_ddtrace: true
+  import_django: true
 import_ddtrace_auto:
   <<: *defaults
   import_ddtrace_auto: true
@@ -28,6 +40,10 @@ import_ddtrace_auto_flask:
   <<: *defaults
   import_ddtrace_auto: true
   import_flask: true
+import_ddtrace_auto_django:
+  <<: *defaults
+  import_ddtrace_auto: true
+  import_django: true
 ddtrace_run_import_ddtrace_auto:
   <<: *defaults
   ddtrace_run: true
@@ -37,3 +53,8 @@ ddtrace_run_import_ddtrace_auto_flask:
   ddtrace_run: true
   import_ddtrace_auto: true
   import_flask: true
+ddtrace_run_import_ddtrace_auto_django:
+  <<: *defaults
+  ddtrace_run: true
+  import_ddtrace_auto: true
+  import_django: true