--- a/benchmarks/startup/scenario.py
+++ b/benchmarks/startup/scenario.py
@@ -1,4 +1,3 @@
-from dataclasses import field
 import os
 import subprocess
 
@@ -9,22 +8,31 @@ class Startup(bm.Scenario):
     ddtrace_run: bool
     import_ddtrace: bool
     import_ddtrace_auto: bool
-    env: dict = field(default_factory=dict)
+    import_flask: bool
+    env: dict[str, str]
 
     def run(self):
         env = os.environ.copy()
         env.update(self.env)
 
-        args = ["python", "-c", ""]
+        commands: list[str] = []
         if self.import_ddtrace:
-            args = ["python", "-c", "import ddtrace"]
-        elif self.import_ddtrace_auto:
-            args = ["python", "-c", "import ddtrace.auto"]
-        elif self.ddtrace_run:
-            args = ["ddtrace-run", "python", "-c", ""]
+            commands.append("import ddtrace")
+        if self.import_ddtrace_auto:
+            commands.append("import ddtrace.auto")
 
-        def _(loops):
+        # Always do this last, we want to import/patch ddtrace first
+        # DEV: We don't need to create/start an app, we just want to see the impact on import time
+        #      with and without patching. We have other benchmarks to test overhead of requests.
+        if self.import_flask:
+            commands.append("import flask")
+
+        args = ["python", "-c"] + [";".join(commands)]
+        if self.ddtrace_run:
+            args = ["ddtrace-run"] + args
+
+        def _(loops: int):
             for _ in range(loops):
-                subprocess.check_call(args, env=env)
+                subprocess.check_call(args=args, env=env)
 
         yield _