@christophe-papazian 2025-02-28
It would be nice to simulate a real prod env. Looking at metabase, I see that many flask users also installed:
flask_cors
flask_sqlalchemy
flask_caching
flask_jwt_extended
flask_babel
flask_migrate
flask_appbuilder
flask_limiter
flask_compress
flask_session

For Django users:
django-cors-headers
django-oauth-toolkit
django-storages
django-appconf
django-filter
django_extensions

For FastAPI:
fastapi_cli
fastapi.utils
fastapi_pagination
fastapi.cli
fastapi_utils

for all users, most used:
certifi
pytz
cryptography
deprecated
botocore
boto3
aiohappyeyeballs
aiohttp
starlette
propcache
sqlalchemy