@brettlangdon 2025-02-28
> It would be nice to simulate a real prod env. Looking at metabase, I see that many flask users also installed: flask_cors flask_sqlalchemy flask_caching flask_jwt_extended flask_babel flask_migrate flask_appbuilder flask_limiter flask_compress flask_session
> 
> For Django users: django-cors-headers django-oauth-toolkit django-storages django-appconf django-filter django_extensions
> 
> For FastAPI: fastapi_cli fastapi.utils fastapi_pagination fastapi.cli fastapi_utils
> 
> for all users, most used: certifi pytz cryptography deprecated botocore boto3 aiohappyeyeballs aiohttp starlette propcache sqlalchemy

@christophe-papazian thinking more about this, we might just want to create dedicated tests for more realistic apps. I'd like to keep this benchmark as "micro" as we can, especially since we do have the `flask_simple` and `django_simple` benchmarks which do try to replicate an actual app (ok, not a lot of extensions/extras, but at least spin up django with an endpoint and hit it).

Maybe a better follow-up is having a `flask_startup` and `django_startup` type benchmarks instead to focus on the added overhead to various django set ups.

I think that and this benchmark could work together. e.g. if we see an increase in this benchmark, we'd expect to see an increase in the flask and django scenarios... or if we see an increase in the django startup benchmark, but not this one, then it helps narrow the problem.



Not 100% sold in either direction, cool with iterating as a follow-up?