@brettlangdon 2025-02-28
Note the microbenchmark for this new benchmark is failing in CI because the scenario technically exists on the baseline, but it is missing a `config.yaml`, so it just fails with that file missing.

> FileNotFoundError: [Errno 2] No such file or directory: 'config.yaml'

I'll investigate, not enabling the benchmark in CI during this PR and doing that as a follow-up might be the better option. I'll keep as enabled in CI for now so any iteration we can validate that it will execute properly on the candidate branch.