@brettlangdon 2025-02-28
> > viztracer supports tracing into subprocesses, but it will only inject into processes which start with python, so the ddtrace-run variants won't get traced.
> 
> This is added in [gaogaotiantian/viztracer#562](https://github.com/gaogaotiantian/viztracer/pull/562). Not sure when the next release will be but the feature will be there.

hey @gaogaotiantian ! thanks for responding, and thanks for `viztracer` !